backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Pydantic
pydantic==2.9.2
pydantic-settings==2.6.1

# LLM
litellm==1.52.16

# Templating
jinja2==3.1.4

# Observability (optional - Arize Phoenix)
arize-phoenix==5.5.0
openinference-instrumentation-litellm==0.1.27

# Utilities
python-dotenv==1.0.1
numpy==2.1.3

# Document processing
pypdf==5.1.0
python-docx==1.1.2

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
